+++ experiments/run_v16_model_comparison.py
@@ -122,7 +122,7 @@ def run_single_model(
     num_soft_tokens = 8
     d_out = num_soft_tokens * embed_dim
     W = make_row_orthonormal_W(d_latent, d_out, seed=1234)
-    W = W.to(device=encoder._device, dtype=encoder.model.dtype)
+    W = W.to(device=encoder._device)  # Keep float32 for precision
     print(f"W shape: {W.shape}, device: {W.device}", flush=True)
 
     # Tasks
